Extract admission controller and call records into admission/controller.rs
